Start work on Android BLE
Beginning of Android code to communicate with the micro:bit over BLE. All there in theory; does not currently work, but does not crash. Seems to be an issue with the way the code attempts to find the micro:bit to connect to, rather than an issue with it attempting to connect over Bluetooth.