Bluetooth (#8)
###
Added Bluetooth functionality to main.cpp
Developed the app to connect to the micro:bit over BLE and read accelerometer data from it.
###

* Started to set up Bluetooth code

Added functions to be used by event listeners. main() function still needs to add a listener for Bluetooth functionality.

* Additional (very early stages) bluetooth code.

More micro:bit code for Bluetooth listeners, very beginnings of Android code.

* UART + select BLE services

Includes working UART example, largely with credit to https://lancaster-university.github.io/microbit-docs/ble/uart-service/#example-microbit-application-animal-vegetable-mineral-game
Added `config.json`; of particular note is the `gatt_table_size` being set to the recommended max of `0x700` and `nested_heap_proportion` being set to `0.50`.
Added `run.bat` to speed up testing without the need for an IDE.

* BLE Accelerometer service C++ working

Fixed the BLE Accelerometer service in main.cpp. It is important to note that as of 06/08/2018 v2.0.0-rc9 of the microbit-DAL for the service to work correctly, though this may be updated.

* Fixed stashing issues

Removed extra text in main.cpp and config.json caused by an earlier stash.

* Fixed stashing issues

Removed extra text in main.cpp caused by an earlier stash.

* Start work on Android BLE

Beginning of Android code to communicate with the micro:bit over BLE. All there in theory; does not currently work, but does not crash. Seems to be an issue with the way the code attempts to find the micro:bit to connect to, rather than an issue with it attempting to connect over Bluetooth.

* Android connects to micro:bit

Finished BLE code to connect to a (pre-paired) micro:bit with a specific MAC address, and read data from the services with the specified UUIDs. Also handles necessary permissions. Currently, data read over BLE is simply outputted to the console. Also includes minor design changes.

* Fully commented MainActivity.java

Added additional comments to MainActivity.java, as well as made some minor formatting changes.

* Process BLE Accelerometer data

App now converts the received BLE accelerometer data from Little Endian byte array to a signed short, and logs the result.

* UI change and displays accelerometer data

Overhauled the UI to a single screen. Now displays the X, Y and Z accelerometer values in the Main Activity as they are received over BLE.

* Integrated Bluetooth with existing micro:bit functionality

Combined C++ programs to include BLE in the main C++ code. Added a "Connection Mode" that pauses vibration checking and waits for the app to connect over BLE.