Stores calibration results in memory
micro:bit now uses uBit.storage to store the 'off' x and y values, and the threshold, in persistent memory. These values are stored during calibration, and loaded from memory at boot. (assuming it is not the first boot since flash)

Added feature to re-calibrate using Button A.
Added copyFromMemory and storeValue functions for general usage of the micro:bit's persistent memory.
Added dictionary of default values for x, y and threshold, and resetDefaultValue function to reset a variable to its default value. This can be expanded and used later, particularly if an invalid value is attempted over BLE and the program has to resort to the default value for the variable to ensure the program can continue to run.