Fix BLE service discovery
Fixed an issue seemingly caused by Android 8.1, in which only generic services could be initially found when calling BluetoothGatt.discoverServices. This is because Android only requests the device once to discover its services, and all subsequent calls to discoverServices simply fetch the cached results from the first call. To fix this, if these services are not found when the code reaches onServicesDiscovered, the program uses reflection to call the inaccessible BluetoothGatt.refresh method. This forces the cache to be cleared and a full rediscovery of all BLE services to occur, causing the non-generic services to be found if they weren't previously.