Added BLE scanner and device select activity
Added a new activity to scan for available BLE devices, and allow the user to select the one they wish to connect to. Highlights likely micro:bit devices in green. Passes the MAC address to the activity that is launched next. Includes a BLEListAdapter class to store discovered BluetoothDevices and to display them in the ListView.