Finalise micro:bit program and add settings cancel
After testing, disallowed only 1ms as an accelerometer period; this one causes memory issues, all others are fine.
Due to readUntil() blocking until data is received (no way around this due to the microbit library) added a Cancel feature in the app, as an unavoidable workaround. If the micro:bit receives this, it will stop listening (as at that point data was received so readUntil() exits) and not update settings.